Updated main code with V3 version
